smoke test of examples
