+++ tests/regression/test_quasistatic.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass
 from conmech.problem_solver import Quasistatic as QuasistaticProblem
 from conmech.problems import Quasistatic
 from examples.p_slope_contact_law import make_slope_contact_law
-from tests.examples_regression.std_boundary import standard_boundary_nodes
+from tests.regression.std_boundary import standard_boundary_nodes
 
 
 @pytest.fixture(
