+++ tests/regression/test_static.py
@@ -1,23 +1,18 @@
 """
 Created at 21.08.2019
 """
+from dataclasses import dataclass
 
 import numpy as np
 import pytest
-from dataclasses import dataclass
 
-from conmech.problem_solver import Dynamic as DynamicProblem
-from conmech.problems import Dynamic
+from conmech.problem_solver import Static as StaticProblem
+from conmech.problems import Static
 from examples.p_slope_contact_law import make_slope_contact_law
-from tests.examples_regression.std_boundary import standard_boundary_nodes
+from tests.regression.std_boundary import standard_boundary_nodes
 
 
-@pytest.fixture(
-    params=[  # TODO #28
-        "global optimization",
-        "schur"
-    ]
-)
+@pytest.fixture(params=["direct", "global optimization", "schur"])
 def solving_method(request):
     return request.param
 
@@ -28,15 +23,12 @@ def generate_test_suits():
     # Simple example
 
     @dataclass()
-    class DynamicSetup(Dynamic):
+    class StaticSetup(Static):
         grid_height: ... = 1
         elements_number: ... = (2, 5)
         mu_coef: ... = 4
         la_coef: ... = 4
-        th_coef: ... = 4
-        ze_coef: ... = 4
-        time_step: ... = 0.1
-        contact_law: ... = make_slope_contact_law(slope=1e1)
+        contact_law: ... = make_slope_contact_law(slope=1)
 
         @staticmethod
         def inner_forces(x, y):
@@ -58,21 +50,21 @@ def generate_test_suits():
         def is_dirichlet(x):
             return x[0] == 0
 
-    setup_m02_m02 = DynamicSetup()
+    setup_m02_m02 = StaticSetup()
 
     expected_displacement_vector_m02_m02 = [
         [0.0, 0.0],
-        [0.03194299, 0.02783408],
-        [0.04913394, 0.04305829],
-        [0.05786617, 0.05507215],
-        [0.06119214, 0.06324329],
-        [0.06012923, 0.07002092],
-        [0.05260607, 0.077619],
-        [0.04304498, 0.08011945],
-        [0.04043799, 0.06935685],
-        [0.03238124, 0.05635081],
-        [0.01987252, 0.0388462],
-        [0.00722258, 0.01704675],
+        [0.04843176, 0.04429129],
+        [0.07642598, 0.08750102],
+        [0.09064929, 0.13301396],
+        [0.09615189, 0.17619656],
+        [0.09566479, 0.21772099],
+        [0.05373533, 0.22480478],
+        [0.00962312, 0.22716041],
+        [0.00730758, 0.18136726],
+        [0.00104241, 0.13297417],
+        [-0.00636696, 0.081923],
+        [-0.00867296, 0.03338082],
         [0.0, 0.0],
         [0.0, 0.0]
     ]
@@ -81,7 +73,7 @@ def generate_test_suits():
 
     # p = 0 and opposite forces
 
-    setup_0_02_p_0 = DynamicSetup()
+    setup_0_02_p_0 = StaticSetup()
     setup_0_02_p_0.contact_law = make_slope_contact_law(slope=0)
 
     def inner_forces(x, y):
@@ -91,17 +83,17 @@ def generate_test_suits():
 
     expected_displacement_vector_0_02_p_0 = [
         [0.0, 0.0],
-        [-0.10204928, -0.11029064],
-        [-0.16607854, -0.27796998],
-        [-0.19804137, -0.48509912],
-        [-0.20984607, -0.70344499],
-        [-0.21187304, -0.91900424],
-        [0.00000004, -0.91885112],
-        [0.21187312, -0.91900425],
-        [0.20984614, -0.70344499],
-        [0.19804144, -0.48509913],
-        [0.16607859, -0.27796999],
-        [0.10204931, -0.11029065],
+        [-0.11787841, -0.1252073],
+        [-0.19322311, -0.31878912],
+        [-0.23173236, -0.56035602],
+        [-0.24636923, -0.8166048],
+        [-0.24898703, -1.07004206],
+        [0.0, -1.06985119],
+        [0.24898703, -1.07004206],
+        [0.24636923, -0.8166048],
+        [0.23173236, -0.56035602],
+        [0.19322311, -0.31878912],
+        [0.11787841, -0.1252073],
         [0.0, 0.0],
         [0.0, 0.0]
     ]
@@ -111,7 +103,7 @@ def generate_test_suits():
 
     # p = 0
 
-    setup_0_m02_p_0 = DynamicSetup()
+    setup_0_m02_p_0 = StaticSetup()
     setup_0_m02_p_0.contact_law = make_slope_contact_law(slope=0)
 
     def inner_forces(x, y):
@@ -129,14 +121,11 @@ def generate_test_suits():
     # various changes
 
     @dataclass()
-    class DynamicSetup(Dynamic):
+    class StaticSetup(Static):
         grid_height: ... = 1.37
         elements_number: ... = (2, 5)
         mu_coef: ... = 4.58
         la_coef: ... = 3.33
-        th_coef: ... = 2.11
-        ze_coef: ... = 4.99
-        time_step: ... = 0.1
         contact_law: ... = make_slope_contact_law(slope=2.71)
 
         @staticmethod
@@ -159,20 +148,20 @@ def generate_test_suits():
         def is_dirichlet(x):
             return x[0] == 0
 
-    setup_var = DynamicSetup()
+    setup_var = StaticSetup()
     expected_displacement_vector_var = [
         [0.0, 0.0],
-        [0.00805403, 0.04399692],
-        [0.00357147, 0.13577263],
-        [-0.01047404, 0.26004496],
-        [-0.0274148, 0.40577722],
-        [-0.04940713, 0.56621234],
-        [-0.21158738, 0.59684774],
-        [-0.40204546, 0.6157323],
-        [-0.36922423, 0.44259321],
-        [-0.31434358, 0.29464635],
-        [-0.24006106, 0.1695022],
-        [-0.14068317, 0.07419011],
+        [-0.02154956, 0.01364313],
+        [-0.04849654, 0.05059958],
+        [-0.07590132, 0.0972985],
+        [-0.09873572, 0.15498692],
+        [-0.12252541, 0.22719522],
+        [-0.19937449, 0.26118308],
+        [-0.30552747, 0.28092124],
+        [-0.27474735, 0.1939756],
+        [-0.22880436, 0.13188258],
+        [-0.17312159, 0.08296667],
+        [-0.10282189, 0.04289061],
         [0.0, 0.0],
         [0.0, 0.0]
     ]
@@ -183,15 +172,11 @@ def generate_test_suits():
 
 
 @pytest.mark.parametrize("setup, expected_displacement_vector", generate_test_suits())
-def test_global_optimization_solver(
-    solving_method, setup, expected_displacement_vector
-):
-    runner = DynamicProblem(setup, solving_method)
-    results = runner.solve(n_steps=32,
-                           initial_displacement=setup.initial_displacement,
-                           initial_velocity=setup.initial_velocity)
+def test_direct_solver(solving_method, setup, expected_displacement_vector):
+    runner = StaticProblem(setup, solving_method)
+    result = runner.solve(initial_displacement=setup.initial_displacement)
 
-    displacement = results[-1].mesh.initial_nodes[:] - results[-1].displaced_points[:]
+    displacement = result.mesh.initial_nodes[:] - result.displaced_points[:]
     std_ids = standard_boundary_nodes(runner.mesh.initial_nodes, runner.mesh.elements)
 
     # print result
